+++ platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/location/LocationComponent.java
@@ -45,6 +45,7 @@ import static android.Manifest.permission.ACCESS_COARSE_LOCATION;
 import static android.Manifest.permission.ACCESS_FINE_LOCATION;
 import static com.mapbox.mapboxsdk.location.LocationComponentConstants.DEFAULT_FASTEST_INTERVAL_MILLIS;
 import static com.mapbox.mapboxsdk.location.LocationComponentConstants.DEFAULT_INTERVAL_MILLIS;
+import static com.mapbox.mapboxsdk.location.LocationComponentConstants.DEFAULT_TRACKING_PADDING_ANIM_DURATION;
 import static com.mapbox.mapboxsdk.location.LocationComponentConstants.DEFAULT_TRACKING_TILT_ANIM_DURATION;
 import static com.mapbox.mapboxsdk.location.LocationComponentConstants.DEFAULT_TRACKING_ZOOM_ANIM_DURATION;
 import static com.mapbox.mapboxsdk.location.LocationComponentConstants.TRANSITION_ANIMATION_DURATION_MS;
@@ -596,6 +597,84 @@ public final class LocationComponent {
   public void cancelZoomWhileTrackingAnimation() {
     checkActivationState();
     locationAnimatorCoordinator.cancelZoomAnimation();
+  }
+
+  /**
+   * Sets the padding.
+   * This API can only be used in pair with camera modes other than {@link CameraMode#NONE}.
+   * If you are not using any of {@link CameraMode} modes,
+   * use one of {@link MapboxMap#moveCamera(CameraUpdate)},
+   * {@link MapboxMap#easeCamera(CameraUpdate)} or {@link MapboxMap#animateCamera(CameraUpdate)} instead.
+   * <p>
+   * If the camera is transitioning when the padding change is requested, the call is going to be ignored.
+   * Use {@link CameraTransitionListener} to chain the animations, or provide the padding as a camera change argument.
+   * </p>
+   *
+   * @param padding The desired padding.
+   */
+  public void paddingWhileTracking(double[] padding) {
+    paddingWhileTracking(padding, DEFAULT_TRACKING_PADDING_ANIM_DURATION, null);
+  }
+
+  /**
+   * Sets the padding.
+   * This API can only be used in pair with camera modes other than {@link CameraMode#NONE}.
+   * If you are not using any of {@link CameraMode} modes,
+   * use one of {@link MapboxMap#moveCamera(CameraUpdate)},
+   * {@link MapboxMap#easeCamera(CameraUpdate)} or {@link MapboxMap#animateCamera(CameraUpdate)} instead.
+   * <p>
+   * If the camera is transitioning when the padding change is requested, the call is going to be ignored.
+   * Use {@link CameraTransitionListener} to chain the animations, or provide the padding as a camera change argument.
+   * </p>
+   *
+   * @param padding           The desired padding.
+   * @param animationDuration The padding animation duration.
+   */
+  public void paddingWhileTracking(double[] padding, long animationDuration) {
+    paddingWhileTracking(padding, animationDuration, null);
+  }
+
+  /**
+   * Sets the padding.
+   * This API can only be used in pair with camera modes other than {@link CameraMode#NONE}.
+   * If you are not using any of {@link CameraMode} modes,
+   * use one of {@link MapboxMap#moveCamera(CameraUpdate)},
+   * {@link MapboxMap#easeCamera(CameraUpdate)} or {@link MapboxMap#animateCamera(CameraUpdate)} instead.
+   * <p>
+   * If the camera is transitioning when the padding change is requested, the call is going to be ignored.
+   * Use {@link CameraTransitionListener} to chain the animations, or provide the padding as a camera change argument.
+   * </p>
+   *
+   * @param padding           The desired padding.
+   * @param animationDuration The padding animation duration.
+   * @param callback          The callback with finish/cancel information
+   */
+  public void paddingWhileTracking(double[] padding, long animationDuration,
+                                   @Nullable MapboxMap.CancelableCallback callback) {
+    checkActivationState();
+    if (!isLayerReady) {
+      notifyUnsuccessfulCameraOperation(callback, null);
+      return;
+    } else if (getCameraMode() == CameraMode.NONE) {
+      notifyUnsuccessfulCameraOperation(callback, String.format("%s%s",
+              "LocationComponent#paddingWhileTracking method can only be used",
+              " when a camera mode other than CameraMode#NONE is engaged."));
+      return;
+    } else if (locationCameraController.isTransitioning()) {
+      notifyUnsuccessfulCameraOperation(callback,
+              "LocationComponent#paddingWhileTracking method call is ignored because the camera mode is transitioning");
+      return;
+    }
+
+    locationAnimatorCoordinator.feedNewPadding(padding, mapboxMap.getCameraPosition(), animationDuration, callback);
+  }
+
+  /**
+   * Cancels animation started by {@link #paddingWhileTracking(double[], long, MapboxMap.CancelableCallback)}.
+   */
+  public void cancelPaddingWhileTrackingAnimation() {
+    checkActivationState();
+    locationAnimatorCoordinator.cancelPaddingAnimation();
   }
 
   /**